requirements.txt
# VoIQ - Vocabulary Intelligence Quiz System
# Python Dependencies

# UI Framework
gradio>=4.0.0

# AI Agent Framework
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-groq>=0.2.0

# Utilities
python-dotenv>=1.0.0
openpyxl>=3.1.0      # Excel file reading (Python fallback)

# Rust Build Tool
maturin>=1.4.0

# Database (Python fallback until Rust is ready)
# SQLite is built into Python

# Testing
pytest>=8.0.0
